import: fix SUDOC authors facette
* Adds `_text` for SUDOC contribubtion with $ref.
